backend/requirements.txt
fastapi
uvicorn
openai-whisper
yt-dlp
llama-index
llama-index-embeddings-huggingface
chromadb
sentence-transformers
python-multipart